Fix comprehensive codegen bugs in Seraphim compiler
Major C code generation fixes:

1. Match expressions: Generate nested ternary operators instead of
   invalid if-statement expressions. Match is now a proper value-
   producing expression: ({ __auto_type __tmp = x; (__tmp == 1) ?
   (10) : ((__tmp == 2) ? (20) : (0)); })

2. Pointer types: Added AST_TYPE_POINTER handling alongside REF types.
   Pointer types now correctly emit T* instead of void.

3. Address-of operator: Added SERAPH_TOK_BIT_AND to unary operators
   for proper &expr code generation.

4. main() return type: C standard requires main() to return int.
   Forward declarations, function definitions, and return statements
   now handle main specially with (int) casts.

5. Enum forward declarations: Added typedef enum X X; to forward
   declarations so enum types can be used before definition.

6. Enum variants: Removed type name prefix from variants. Seraph
   enums use global variant names (TokIf, not TokenType_TokIf).

7. Named types: Removed automatic 'struct' prefix for AST_TYPE_NAMED.
   Types now emit just the typedef name, fixing enum parameter usage.

8. Pointer auto-deref: Added tracking of pointer identifiers to
   generate -> instead of . for field access on pointers, matching
   Seraph/Rust semantics where ptr.field auto-dereferences.

9. str type: Added mapping from 'str' to 'const char*' for string
   type support.

All compiler integration tests pass.

+++ tests/array_test_c.c
@@ -69,11 +69,11 @@ typedef struct { int64_t real; int64_t dual; } Seraph_Dual;
 typedef struct { int64_t w; int64_t x; int64_t y; int64_t z; } Seraph_Galactic;
 
 /* Forward declarations */
-int32_t main(void);
+int main(void);
 
-int32_t main(void) {
+int main(void) {
     int32_t arr[4] = { 10LL, 20LL, 30LL, 40LL };
     int32_t x = arr[0LL];
     int32_t y = arr[2LL];
-    return (x + y);
+    return (int)(x + y);
 }
