+++ tests/fnptr_callonly_c.c
@@ -72,7 +72,7 @@ typedef struct { int64_t w; int64_t x; int64_t y; int64_t z; } Seraph_Galactic;
 int32_t a(void);
 int32_t b(void);
 int32_t invoke(int32_t (*f)(void));
-int32_t main(void);
+int main(void);
 
 int32_t a(void) {
     return 100LL;
@@ -83,8 +83,8 @@ int32_t b(void) {
 int32_t invoke(int32_t (*f)(void)) {
     return f();
 }
-int32_t main(void) {
+int main(void) {
     invoke(a);
     invoke(b);
-    return invoke(a);
+    return (int)invoke(a);
 }
