+++ tests/fnptr_which_c.c
@@ -72,7 +72,7 @@ typedef struct { int64_t w; int64_t x; int64_t y; int64_t z; } Seraph_Galactic;
 int32_t ret_one(void);
 int32_t ret_two(void);
 int32_t call_it(int32_t (*f)(void));
-int32_t main(void);
+int main(void);
 
 int32_t ret_one(void) {
     return 1LL;
@@ -83,6 +83,6 @@ int32_t ret_two(void) {
 int32_t call_it(int32_t (*f)(void)) {
     return f();
 }
-int32_t main(void) {
-    return call_it(ret_one);
+int main(void) {
+    return (int)call_it(ret_one);
 }
