+++ tests/while_test_c.c
@@ -69,14 +69,14 @@ typedef struct { int64_t real; int64_t dual; } Seraph_Dual;
 typedef struct { int64_t w; int64_t x; int64_t y; int64_t z; } Seraph_Galactic;
 
 /* Forward declarations */
-int32_t main(void);
+int main(void);
 
-int32_t main(void) {
+int main(void) {
     int32_t sum = 0LL;
     int32_t i = 0LL;
     while ((i < 10LL)) {
         (sum = (sum + i));
         (i = (i + 1LL));
     }
-    return sum;
+    return (int)sum;
 }
